Moved profile info card components to components folder and fixed styling
